# Add an expiring cache for the caching token authenticator