--- a/staging/src/k8s.io/apiserver/pkg/util/cache/BUILD
+++ b/staging/src/k8s.io/apiserver/pkg/util/cache/BUILD
@@ -1,0 +1,13 @@
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)