… staging/src/k8s.io/cloud-provider/go.sum
… staging/src/k8s.io/kubectl/go.sum
… staging/src/k8s.io/node-api/go.sum
… staging/src/k8s.io/sample-controller/go.sum