@mikedanese 2019-10-29
I played around with:

* Making cleanup use persistent goroutine
* Making cleanup batch deletes.
* Not performing heap operations with exclusive lock.

This persistent goroutine reduces garbage by about 60%. This was 350 MB -> 150 MB over 10 seconds under the GC stress test. The other two appear to be mostly negligible at the benchmark's read/write ratio.

Overall diff:
```
name           old time/op    new time/op    delta
ExpiringCache     574ns ± 3%     456ns ± 4%  -20.54%  (p=0.000 n=8+10)
name           old alloc/op   new alloc/op   delta
ExpiringCache     9.00B ± 0%     4.00B ± 0%  -55.56%  (p=0.002 n=8+10)
name           old allocs/op  new allocs/op  delta
ExpiringCache      0.00           0.00          ~     (all equal)
```