@mikedanese 2019-11-08
this is ready for another review