@mikedanese 2019-11-13
Preliminary benchmark shows that the striping still helps out a bit. Here's a before/after benchmark for the last revision:

```
name                                                  old time/op     new time/op     delta
CachedTokenAuthenticator/tokens=100_threads=1-12         1.13µs ± 3%     1.11µs ± 7%      ~     (p=0.839 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=16-12        1.53µs ± 6%     1.08µs ± 1%   -29.39%  (p=0.000 n=9+7)
CachedTokenAuthenticator/tokens=100_threads=256-12       2.71µs ±12%     1.07µs ± 6%   -60.51%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=500_threads=1-12         1.15µs ± 3%     1.06µs ± 6%    -7.55%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=16-12        1.76µs ± 3%     1.05µs ± 4%   -40.40%  (p=0.000 n=9+10)
CachedTokenAuthenticator/tokens=500_threads=256-12       3.03µs ±19%     1.15µs ± 4%   -61.88%  (p=0.000 n=10+9)
CachedTokenAuthenticator/tokens=2500_threads=1-12        1.17µs ± 1%     1.09µs ± 1%    -7.34%  (p=0.000 n=9+9)
CachedTokenAuthenticator/tokens=2500_threads=16-12       2.02µs ± 5%     1.00µs ± 2%   -50.34%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=256-12      2.65µs ± 9%     1.17µs ± 2%   -56.00%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-12       1.79µs ± 2%     1.67µs ± 1%    -6.64%  (p=0.000 n=10+8)
CachedTokenAuthenticator/tokens=12500_threads=16-12      2.19µs ±10%     1.12µs ± 3%   -48.96%  (p=0.000 n=9+10)
CachedTokenAuthenticator/tokens=12500_threads=256-12     2.43µs ± 8%     1.29µs ± 3%   -47.08%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=1-12       53.0µs ± 0%     43.3µs ± 0%   -18.39%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=16-12      13.8µs ± 2%      1.7µs ± 3%   -87.42%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-12     14.0µs ± 1%      2.3µs ± 6%   -83.39%  (p=0.000 n=10+9)

name                                                  old alloc/op    new alloc/op    delta
CachedTokenAuthenticator/tokens=100_threads=1-12           327B ± 0%       295B ± 0%    -9.79%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=16-12          332B ± 0%       299B ± 0%    -9.94%  (p=0.000 n=9+10)
CachedTokenAuthenticator/tokens=100_threads=256-12         300B ± 0%       267B ± 0%   -11.13%  (p=0.000 n=7+10)
CachedTokenAuthenticator/tokens=500_threads=1-12           329B ± 0%       296B ± 0%   -10.03%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=16-12          297B ± 0%       296B ± 0%    -0.34%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=256-12         332B ± 0%       301B ± 0%    -9.39%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=1-12          328B ± 0%       295B ± 0%   -10.06%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=16-12         329B ± 0%       264B ± 0%   -19.68%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=256-12        301B ± 0%       300B ± 0%    -0.23%  (p=0.002 n=6+10)
CachedTokenAuthenticator/tokens=12500_threads=1-12         330B ± 0%       297B ± 0%   -10.00%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=16-12        331B ± 0%       265B ± 0%   -19.94%  (p=0.002 n=8+10)
CachedTokenAuthenticator/tokens=12500_threads=256-12       336B ± 0%       270B ± 0%   -19.52%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=1-12         469B ± 0%       371B ± 0%   -20.90%  (p=0.000 n=9+9)
CachedTokenAuthenticator/tokens=62500_threads=16-12        408B ± 0%       273B ± 0%   -33.01%  (p=0.000 n=9+10)
CachedTokenAuthenticator/tokens=62500_threads=256-12       427B ± 0%       313B ± 1%   -26.58%  (p=0.000 n=10+10)

name                                                  old allocs/op   new allocs/op   delta
CachedTokenAuthenticator/tokens=100_threads=1-12           13.0 ± 0%       11.0 ± 0%   -15.38%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=16-12          13.0 ± 0%       11.0 ± 0%   -15.38%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=256-12         12.0 ± 0%       10.0 ± 0%   -16.67%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=1-12           13.0 ± 0%       11.0 ± 0%   -15.38%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=16-12          12.0 ± 0%       11.0 ± 0%    -8.33%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=256-12         13.0 ± 0%       11.0 ± 0%   -15.38%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=1-12          13.0 ± 0%       11.0 ± 0%   -15.38%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=16-12         13.0 ± 0%       10.0 ± 0%   -23.08%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=256-12        12.0 ± 0%       11.0 ± 0%    -8.33%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-12         13.0 ± 0%       11.0 ± 0%   -15.38%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=16-12        13.0 ± 0%       10.0 ± 0%   -23.08%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=256-12       13.0 ± 0%       10.0 ± 0%   -23.08%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=1-12         17.0 ± 0%       13.0 ± 0%   -23.53%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=16-12        16.0 ± 0%       10.0 ± 0%   -37.50%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-12       16.0 ± 0%       12.0 ± 0%   -25.00%  (p=0.000 n=10+10)
```