@lavalamp 2019-11-13
It's weird that "CachedTokenAuthenticator/tokens=62500_threads=1-12" remains quite slow compared with the runs with more threads which were drastically improved. Why is that, any ideas?