@mikedanese 2019-11-13
> It's weird that "CachedTokenAuthenticator/tokens=62500_threads=1-12" remains quite slow compared with the runs with more threads which were drastically improved. Why is that, any ideas?

Each load has a fixed cost due to the sleep in the benchmark. In the single threaded case, those loads happen serially and at 62500 tokens, it dominates the runtime of the benchmark. Here's the raw results for a 10s benchtime:

```
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=1-12               129572             93308 ns/op             434 B/op         16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=1-12               129007             92894 ns/op             434 B/op         16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=1-12               129566             92662 ns/op             434 B/op         16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=1-12               129859             92936 ns/op             434 B/op         16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=1-12               129091             92832 ns/op             434 B/op         16 allocs/op
```

On larger thread numbers, that cost is distributed over multiple threads and happens in parallel and after the cache is loaded, things run quickly.